migration/scripts/requirements.txt
# Climasys Migration Script Dependencies
# Install with: pip install -r requirements.txt

# Database connectivity
psycopg2-binary>=2.9.0

# Data processing
pandas>=1.5.0

# Environment configuration
python-dotenv>=0.19.0

# Optional: For enhanced CSV processing
openpyxl>=3.0.0

# Optional: For better logging
colorlog>=6.0.0
